rename all component class name and file name to have the component suffix
