+++ components/shooterComponent.py
